refactor(members): move members into the settings page
The sidebar was already scrolling before this work, and adding a top-level
Members entry made a long list longer for a screen that is organization
configuration rather than a product module.

Members now render as a card on /dashboard/settings beneath the organization
settings card, behind the same canManageOrgSettings gate. No new nav row, so
the overflow does not get worse, and it sits where organization configuration
already lives.

Removes the standalone route and the plumbing that only existed to serve it:
the loading skeleton, the DashboardLoadingRoute entry, the shell title and
skeleton cases, and the members arm of CompactOperationsCardSkeleton's route
union, which now has a single member. /members redirects to the settings page
so an existing link still lands somewhere sensible.
