--- apps/sdp-web/src/app/dashboard/members/loading.tsx
@@ -1,5 +1,0 @@
-import { CompactOperationsCardSkeleton } from "../operations-card-page-skeletons";
-
-export default function MembersLoading() {
-  return <CompactOperationsCardSkeleton route="members" />;
-}
