+++ apps/sdp-web/src/app/members/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
 
 export default function MembersPage() {
-  redirect("/dashboard/members");
+  redirect("/dashboard/settings");
 }
