+++ apps/sdp-web/src/lib/dashboard-navigation-loading.ts
@@ -6,7 +6,6 @@ export const DASHBOARD_SIDE_NAV_HREFS = {
   apiKeys: "/dashboard/api-keys",
   policies: "/dashboard/policies",
   approvals: "/dashboard/approvals",
-  members: "/dashboard/members",
   settings: "/dashboard/settings",
 } as const;
 
@@ -48,7 +47,6 @@ export type DashboardLoadingRoute =
   | "policies"
   | "approvals-list"
   | "approval-detail"
-  | "members"
   | "settings"
   | "allowlist";
 
@@ -112,7 +110,6 @@ export function resolveDashboardLoadingRoute(rawPathname: string): DashboardLoad
   if (pathname === "/dashboard/policies") return "policies";
   if (pathname === "/dashboard/approvals") return "approvals-list";
   if (/^\/dashboard\/approvals\/[^/]+$/.test(pathname)) return "approval-detail";
-  if (pathname === "/dashboard/members") return "members";
   if (pathname === "/dashboard/settings") return "settings";
   if (pathname === "/dashboard/allowlist") return "allowlist";
